Split activity page tests into their own file
It’s not intuitive for these tests to be lumped in with the tests for
the job page.
